edf e rm

+++ RM.cpp
@@ -3,7 +3,7 @@
 ProcessParams *RM::select_process(vector<ProcessParams *> processes)
 {
 
-    size_t index_of_priority_process = -1;
+    int index_of_priority_process = -1;
     int highest_priority = 0;
 
     for (size_t i = 0; i < processes.size(); i++)
